fix: things that should have been pushed with the last commit :/

+++ stage.sh
@@ -5,8 +5,12 @@ read api
 echo "Deploy App Changes? (y/n)"
 read app
 if [ "$api" = "y" ] ; then
-    echo "Compiling TypeScript to JavaScript..."
-    tsc -p ./backend/
+    echo "Compile server before uploading? (y/n)"
+    read compileapi
+    if [ "$compileapi" = "y" ] ; then
+        echo "Compiling TypeScript to JavaScript..."
+        tsc -p ./backend/
+    fi
     echo "Copying files to server..."
     ssh killer-whale 'rm -r /var/www/blog_staging/api/build'
     rsync -r ./backend/build killer-whale:/var/www/blog_staging/api
@@ -18,11 +22,15 @@ fi
 # frontend deployment
 if [ "$app" = "y" ] ; then
     echo "Beginning Frontend deployment."
-    echo "Please set the config to the proper blog endpoint before continuing (api.test.blog.paulgellai.dev)."
-    read -p "Press enter to continue. "
-    cd frontend
-    yarn build --prod
-    cd ..
+     echo "Compile app before uploading? (y/n)"
+    read compileapp
+    if [ "$compileapp" = "y" ] ; then
+        echo "Please set the config to the proper blog endpoint before continuing (api.blog.paulgellai.dev)."
+        read -p "Press enter to continue. "
+        cd frontend
+        yarn build --prod
+        cd ..
+    fi
     echo "Copying files to server..."
     ssh killer-whale 'rm -r /var/www/blog_staging/app/'
     rsync -r ./frontend/dist/ killer-whale:/var/www/blog_staging/app/
